Extract SegmentsView
